need to keep it in JabrefDesktop for now as it has dependencies to the gui package
