+++ src/main/java/net/sf/jabref/gui/desktop/os/Linux.java
@@ -1,7 +1,5 @@
 package net.sf.jabref.gui.desktop.os;
 
-import net.sf.jabref.Globals;
-import net.sf.jabref.JabRefPreferences;
 import net.sf.jabref.external.ExternalFileType;
 import net.sf.jabref.external.ExternalFileTypes;
 
@@ -14,8 +12,7 @@ public class Linux implements NativeDesktop {
     @Override
     public void openFile(String filePath, String fileType) throws IOException {
         ExternalFileType type = ExternalFileTypes.getInstance().getExternalFileTypeByExt(fileType);
-        // TODO: use "xdg-open" instead?
-        String viewer = type == null ? Globals.prefs.get(JabRefPreferences.PSVIEWER) : type.getOpenWith();
+        String viewer = type == null ? "xdg-open" : type.getOpenWith();
         String[] cmdArray = new String[2];
         cmdArray[0] = viewer;
         cmdArray[1] = filePath;
