+++ src/main/java/net/sf/jabref/logic/util/OS.java
@@ -15,8 +15,6 @@
  */
 package net.sf.jabref.logic.util;
 
-import net.sf.jabref.gui.desktop.os.*;
-
 /***
  * Operating system (OS) detection
  */
@@ -27,21 +25,4 @@ public class OS {
     public static final boolean LINUX = OS_NAME.startsWith("linux");
     public static final boolean WINDOWS = OS_NAME.startsWith("win");
     public static final boolean OS_X = OS_NAME.startsWith("mac");
-
-    public static final NativeDesktop ND_LINUX = new Linux();
-    public static final NativeDesktop ND_WINDOWS = new Windows();
-    public static final NativeDesktop ND_MAC = new OSX();
-    public static final NativeDesktop ND_DEFAULT = new DefaultDesktop();
-
-    public static NativeDesktop getNativeDesktop() {
-        if(WINDOWS) {
-            return ND_WINDOWS;
-        } else if(OS_X) {
-            return ND_MAC;
-        } else if(LINUX) {
-            return ND_LINUX;
-        }
-        return ND_DEFAULT;
-    }
-
 }
